+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/MessageManager.java
@@ -517,4 +517,22 @@ public class MessageManager extends BaseManager {
                 jsonValue(methodCall, "messageList")
         );
     }
+
+    public void addMessageReactionExtensions(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.addMessageReactionExtensions(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                jsonValue(methodCall, "message"),
+                jsonValue(methodCall, "list")
+        );
+    }
+
+    public void getMessageListSomeReactionExtensions(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getMessageListSomeReactionExtensions(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                jsonValue(methodCall, "messageList"),
+                jsonValue(methodCall, "list")
+        );
+    }
 }